requirements_prod.txt
appdirs==1.4.3
asgiref==3.2.3
distlib==0.3.0
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.5.0
Markdown==3.2
nltk==3.4.4
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
virtualenv==20.0.2
zipp==2.2.0